==> ipynb/01_preprocessing.ipynb <==
# Preprocessing

1. Organizing data
2. Extracting patches

## 1. Organizing data

### Input files

- ```data/TGGATEs/orig/open_tggates_individual.csv```
- ```data/TGGATEs/orig/open_tggates_pathological_image.csv```
- ```data/TGGATEs/orig/open_tggates_pathology.csv```
- ```data/TGGATEs/orig/tggates_in_vivo_compound_info.txt```
- ```data/TGGATEs/orig/compound_table.db```

You can obtain the data respectively from

- http://togodb.org/db/open_tggates_individual
- http://togodb.org/db/open_tggates_pathological_image
- http://togodb.org/db/open_tggates_pathology
- https://toxico.nibiohn.go.jp/english/datalist.html


### Output files

- ```data/TGGATEs/processed/info.csv```
- ```data/TGGATEs/processed/ft_list.txt```
- ```data/TGGATEs/processed/compound_list.txt```

!pwd

import sys

root = ".."
sys.path.append(root)


import toxreprcnn.preprocess
import os
from collections import defaultdict
from tqdm import tqdm
import openslide
import sqlite3
import pandas as pd
import numpy as np
from collections import defaultdict

root = ".."
orig_dir = f"{root}/data/TGGATEs/orig"
processed_dir = f"{root}/data/TGGATEs/processed"


ind_df = pd.read_csv(
    f"{orig_dir}/open_tggates_individual.csv", encoding="shift_jis")
image_df = pd.read_csv(
    f"{orig_dir}/open_tggates_pathological_image.csv", encoding="shift_jis"
)
path_df = pd.read_csv(
    f"{orig_dir}/open_tggates_pathology.csv", encoding="shift_jis")
compound_df = pd.read_csv(
    f"{orig_dir}/tggates_in_vivo_compound_info.txt", encoding="shift_jis", sep="\t"
)

display(ind_df.head())
display(image_df.head())
display(path_df.head())
display(compound_df.head())


### Set new individual IDs for each rat

ind_df.columns = map(lambda x: x.upper(), ind_df.columns)

image_df = image_df[image_df["ORGAN"] == "Liver"]
path_df = path_df[path_df["ORGAN"] == "Liver"]
compound_df = compound_df[compound_df["ORGAN"] == "Liver"]

image_df["INDV_ID"] = (
    image_df["EXP_ID"] * 1000 + image_df["GROUP_ID"] *
    10 + image_df["INDIVIDUAL_ID"]
)

path_df["INDV_ID"] = (
    path_df["EXP_ID"] * 1000 + path_df["GROUP_ID"] *
    10 + path_df["INDIVIDUAL_ID"]
)

ind_df["COMPOUND_NAME"] = ind_df["COMPOUND_NAME"].where(
    ind_df["COMPOUND_NAME"] != "TNFﾎｱ", "TNFα"
)


### Finding Types

path_df["FINDING_TYPE"].head()

ft_list = list(set(path_df["FINDING_TYPE"].dropna()))
len(ft_list)


TGGATEs liver pathological images contain 65 pathological finding types.

# saving the finding type list
open(f"{processed_dir}/ft_list.txt", "w").write("\n".join(ft_list) + "\n")


Transforming the pathological finding type information into one-hot vectors.

one_hot = pd.get_dummies(
    path_df[["INDV_ID", "FINDING_TYPE"]].dropna(),
    columns=["FINDING_TYPE"],
    prefix="",
    prefix_sep="",
)
one_hot = one_hot.groupby("INDV_ID").max()
one_hot.head()

Merging the pathological finding type information with the individual rat data.

ind_ft_df = pd.merge(ind_df, one_hot, on="INDV_ID", how="left").fillna(0)
ind_ft_df.head()

### Compounds

ind_ft_df["COMPOUND_NAME"].head()


Merging the compound information with the individual rat data.

ind_ft_com_df = pd.merge(
    ind_ft_df, compound_df[["COMPOUND_NAME", "VEHICLE"]], on="COMPOUND_NAME", how="left"
)
ind_ft_com_df.head()

If the dose is zero, then the compound name is changed to the vehicle name.

ind_ft_com_df["COMPOUND_NAME"] = ind_ft_com_df["COMPOUND_NAME"].where(
    ind_ft_com_df["DOSE"] != 0, ind_ft_com_df["VEHICLE"]
)
ind_ft_com_df[ind_ft_com_df["DOSE"] == 0].head()

The list of compounds.

compound_list = list(set(ind_ft_com_df["COMPOUND_NAME"]))
print(*compound_list, sep="\t")


# saving the compound list
open(f"{processed_dir}/compound_list.txt",
     "w").write("\n".join(compound_list) + "\n")


Transforming the compound information into one-hot vectors.

one_hot = pd.get_dummies(
    ind_ft_com_df[["INDV_ID", "COMPOUND_NAME"]], prefix="", prefix_sep=""
)
ind_ft_com_df = pd.merge(ind_ft_df, one_hot, on="INDV_ID", how="left")
ind_ft_com_df.head()

### Images

The file location column includes the location from which you can download the pathological images.

image_df["FILE_LOCATION"].head()


Extracting the file names.

image_df["FILE"] = [f[-1] for f in image_df["FILE_LOCATION"].str.split("/")]
image_df["FILE"].head()


Merging the image file information with the individual rat data.

info_df = pd.merge(
    ind_ft_com_df,
    image_df[["INDV_ID", "FILE", "FILE_LOCATION"]],
    on="INDV_ID",
    how="left",
)
info_df.head()

info_df.to_csv(f"{processed_dir}/info.csv", index=False)


path = f"{root}/data/DrugBank/compound_table.db"

conn = sqlite3.connect(path)
conn.row_factory = sqlite3.Row
cur = conn.cursor()
cur.execute("select * from compound_table")
com_dict = cur.fetchall()

key_list = list(com_dict[0].keys())
key_dict = {k: i for i, k in enumerate(key_list)}

table = [[] for i in range(len(key_dict))]

for cd in com_dict:
    for k in cd.keys():
        table[key_dict[k]].append(cd[k])


df = pd.DataFrame(table, index=key_list).T


com_list = list(set(info_df["COMPOUND_NAME"]))
syn_dict = {}
for cid, syn in df[["compound_id", "synonym"]].values:
    for s in syn.split(";"):
        syn_dict[s.lower()] = cid


dd = defaultdict(list)

for com in com_list:
    if com in syn_dict:
        moa = df.loc[df["compound_id"] == syn_dict[com], "mechanism_of_action"].values[
            0
        ]
        if moa is not None:
            for m in moa.split(";"):
                dd[m].append(com)

for k, v in dd.items():
    print(k, len(v))

#### Extract MoA with no less than 3 compounds

for k, v in dd.items():
    if len(v)>=3:
        print(f"MoA : {k}")
        print(*v)

moa_list = []
for k, v in dd.items():
    if len(v)>=3:
        if k!="Unknown":
            moa_list.append(k)
moa_dict = {k:i for i,k in enumerate(moa_list)}

com_set = set()
for moa in moa_list:
    com_set |= set(dd[moa])

extracted_com_list = list(com_set)
extracted_com_dict = {k:i for i,k in enumerate(extracted_com_list)}

matrix = np.zeros(shape = (len(extracted_com_list), len(moa_list)))
for moa in moa_list:
    for com in dd[moa]:
        matrix[extracted_com_dict[com], moa_dict[moa]] = 1

moa_df = pd.DataFrame(matrix, columns = moa_list, index=extracted_com_list)
moa_df.head()

moa_df.to_csv(f"{processed_dir}/moa.csv")

## Extracting patches

- 100 patch/WSI
- patch : 1024x1024 pixels

files = info_df["FILE"].dropna().values

# patch_dir = "/data/TGGATEs/patches"
# wsi_dir = "/data/TGGATEs/WSI"

# for s in tqdm(files):
#     path = f"{wsi_dir}/{s}"
#     if not os.path.exists(f"{patch_dir}/{s}"):
#         os.system(f"mkdir {patch_dir}/{s}")
#         try:
#             toxreprcnn.preprocess.save_tiles(image_path=path, tile_size=1024, n_tiles=100, tile_dir=f"{patch_dir}/{s}")
#         except KeyboardInterrupt:
#             break
#         except:
#             print("Something is wrong with",s)
