==> ipynb/03_data_split.ipynb <==
# Data Split

## import library

import pickle
import random
from glob import glob

import numpy as np
import pandas as pd
from sklearn.model_selection import GroupKFold
from sklearn.preprocessing import LabelEncoder
from tqdm import tqdm

from toxreprcnn.data_split import RepeatedStratifiedGroupKFold


# Data split

root = ".."

with open(f"{root}/processed_data/210925single_ft_tile_predicts.pickle", "rb") as f:
    labels = pickle.load(f)


print(*sorted(list(labels.keys())), sep="\n")

labelsは各finding typeが存在するWSIから抽出されたtileに関してWSSSで予測されたpathological finding存在確率を格納したデータ

finding typeをkeyとするdictで、各valueはtileのリストと存在確率のペアになっている

labels["Proliferation, oval cell"][0][:5]


labels["Proliferation, oval cell"][1][0][:5]


selected finding types

fts = """Proliferation, bile duct
Ground glass appearance
Increased mitosis
Inclusion body, intracytoplasmic
Deposit, pigment
Single cell necrosis
Vacuolization, cytoplasmic
Swelling"""
fts = fts.split("\n")


tile_set = set()

for ft in fts:
    tile_set |= set(labels[ft][0])
positive_set = set([t.split("/")[-2] for t in tile_set])


tile_root = "/data/TGGATE/tiles"

path_list = glob(f"{tile_root}/*.svs")
svs_path_dict = {path.split("/")[-1]: path for path in path_list}
info = pd.read_csv(f"{root}/../../info/info.csv")
info = info[info["SACRI_PERIOD"].isin(["4 day", "8 day", "15 day", "29 day"])]
whole_set = set(svs_path_dict.keys()) & set(info["FILE"])
positive_set = whole_set & positive_set
rest_set = whole_set - positive_set


positive_list = list(positive_set)

positive_info = []

for wsi in tqdm(positive_list):
    for t in glob(svs_path_dict[wsi] + "/*.tiff"):
        positive_info.append((t, "/".join(t.split("/")[-2:]), t.split("/")[-2]))

df = pd.DataFrame(positive_info, columns=["path", "tilename", "FILE"])


for ft in fts:
    dat = []

    data = list(zip(labels[ft][0], np.concatenate(labels[ft][1])))
    for p, l in data:
        tilename = "/".join(p.split("/")[-2:])
        dat.append([tilename, l[0]])
    # add label information
    df = pd.merge(
        df, pd.DataFrame(dat, columns=["tilename", ft]), on="tilename", how="left"
    )


for ft in fts:
    df.loc[df[ft].isnull(), ft] = 0
df = df[((df[fts] > 0) & (df[fts] < 0.5)).sum(axis=1) == 0]


df


negative_list = random.sample(rest_set, k=len(df["FILE"].unique()))

negative_info = []
for wsi in tqdm(negative_list):
    for t in glob(svs_path_dict[wsi] + "/*.tiff"):
        negative_info.append((t, "/".join(t.split("/")[-2:]), t.split("/")[-2]))
df_neg = pd.DataFrame(negative_info, columns=["path", "tilename", "FILE"])
for ft in fts:
    df_neg[ft] = 0


df = pd.concat([df, df_neg])


df = pd.merge(df, info[["EXP_ID", "GROUP_ID", "FILE"]], on="FILE", how="inner")
df["EG"] = df["EXP_ID"] * 100 + df["GROUP_ID"]

model_train_df = df


## Data Split


rsgf = RepeatedStratifiedGroupKFold(n_splits=5, random_state=123, n_repeats=100)
g = LabelEncoder().fit_transform(info["COMPOUND_NAME"].to_numpy())
y = info[fts].values.astype(np.int64)
X = info[["FILE"]].to_numpy()
fold = np.zeros(len(info), dtype=np.int64)

for i, (_, idx) in enumerate(rsgf.split(X, y, g)):
    fold[idx] = i
fold = LabelEncoder().fit_transform(fold)
for i in range(6):
    print(y[fold == i].sum(axis=0))
info["fold 1"] = fold


train1 = info[info["fold 1"] != 0]


spl = np.zeros(len(train1), dtype=np.int64)
spl[train1["SACRI_PERIOD"].isin(["15 day", "29 day"])] = 1
train1["SP_label"] = spl
train1["EG"] = train1["GROUP_ID"] + train1["EXP_ID"] * 100

g = LabelEncoder().fit_transform(train1["EG"].to_numpy())
y = train1[fts + ["SP_label"]].to_numpy().astype(np.int64)
X = train1[["FILE"]].to_numpy()
fold = np.zeros(len(train1), dtype=np.int64)

for i, (_, idx) in enumerate(rsgf.split(X, y, g)):
    fold[idx] = i
fold = LabelEncoder().fit_transform(fold)

for i in range(5):
    print(y[fold == i].sum(axis=0))
train1["fold 2"] = fold


train2 = train1[train1["fold 2"] != 0]
df_train, df_test = (
    model_train_df[model_train_df["FILE"].isin(train2["FILE"])],
    model_train_df[~model_train_df["FILE"].isin(train2["FILE"])],
)


gkf = GroupKFold(n_splits=4)

df_train = df_train.sample(frac=1.0)
X = df_train["path"].to_numpy()
y = df_train[fts].to_numpy()
g = df_train["EG"].to_numpy()
fold = np.zeros(len(df_train), dtype=np.int64)

for i, (_, idx) in enumerate(gkf.split(X, y, g)):
    fold[idx] = i
df_train["fold"] = fold
for i in range(4):
    print(y[fold == i].sum(axis=0), (fold == i).sum())


out_dir = "../../data/TGGATEs/processed"

df_train.to_csv(f"{out_dir}/train_val_for_model_training.csv", index=False)
df_test.to_csv(f"{out_dir}/test_for_finding_types.csv", index=False)
train2.to_csv(f"{out_dir}/train.csv", index=False)
